create RegenerateCommand
